Avoid focus on floating signature help
Signature help seem to ignore the focussable option when the signature is shown
for an "space" character, if the prev char couldn't show the signature. For
example (python example):
 getattr(foo, )
              ^ here if we type fast enough for the first signature to be shown
we get a focus on the opening parens of the signature help window. Annoyingly
enough, the only way to get rid of the window is using the mouse to change the
focus once again.