fix(nvdash): fix crashing while floating windows are present
So Nvdash closes all windows except the current window to display dashboard.
But if floating windows exist like terminal or startup lazy install exists
it crashes so I close all floating windows normally without force for
floating windows and set its window for the first one which is not floating.