fix(nvcheatsheet): fix unicode highlighting and spacing of nvcheatsheet (#86)
* fix(nvcheatsheet): fix unicode highlighting and spacing of nvcheatsheet

As the nvim_buf_add_highlight and other functions use byte indexing instead
of character indexing, I have fixed the whitespace length to not include the
extra bytes of unicode and also for highlighting I had to get the correct byte
index for both text and space lines of cheatsheet using builtin functions.

Additional testing is required to see all cases are handled.
I have checked for emojis and glyphs working as of now.

fixes #74

* fix(nvcheatsheet): handle title highlights as well

For Now this will handle spacing and highlights of title and space below as well.
But emoji in title is bit tricky to solve for now emojis or glyphs in title should not be allowed.

* fix(nvcheatsheet): fix heading nerd font as well

I have fixed nerd font icon for heading title as well. I have tried to include
emojis but yet they are causing the space highlighting to offset right side.
So emoji support yet.