fix nvim-web-devicons filename in tabufline
It is not correct to have different behaviours in Tabufline and
Statusline regarding the nvim-web-devicons `get_icon` call.

For Statusline, it passes the whole expanded filename using %:t, but for
Tabufline, it uses a regex to pass just the file extension, even though
nvim-web-devicons supports passing everything to find the icon.

To fix this, the regex is removed, passing the whole filename passed
through the `add_fileInfo` function to `get_icon` call.