Feat: Add custom keymap capabilities to theme switcher
In order to enable customizing the keymaps for the new theme switcher,
this commit extracts most of the navigation functions out of
mappings.lua into a new api.lua. This allows the navigation functions to
be publicly exposed as a module which can then be used when setting up
custom keymaps.

Additionally, this commit adds the ability to customize the new theme
switcher via Chadrc with a new `theme_switcher` field. This new field on
the Chadrc configuration allows adding key mappings for both insert and
normal mode.